// src/types/interfaces.d.ts
export interface IClientOptions {
    host: string;
    apiHost?: string;
    authToken?: string;
    jwt?: string;
    jid?: string;
    jidResource?: string;
    reconnectOnNoLongerSubscribed?: boolean;
    optOutOfWebrtcStatsTelemetry?: boolean;
    allowIPv6?: boolean;
    logger?: ILogger;
    logLevel?: LogLevel;
    logFormatters?: LogFormatterFn[];
    signalIceConnected?: boolean;
    useServerSidePings?: boolean;
    appName?: string;
    appVersion?: string;
    appId?: string;
    customHeaders?: ICustomHeader;
    /** Allow bulk topic resubscribe to succeed or fail per-topic rather than all or nothing */
    enablePartialBulkResubscribe?: boolean;
    /** Genesys internal use only - non-Genesys apps that pass in `alertableInteractionTypes` may experience unexpected behavior */
    alertableInteractionTypes?: AlertableInteractionTypes[];
}
export interface ICustomHeader {
    [header: string]: string;
}
export declare enum AlertableInteractionTypes {
    voice = "voice"
}
// ...
export interface IClientConfig {
    host: string;
    apiHost: string;
    authToken?: string;
    jwt?: string;
    userId?: string;
    jid?: string;
    jidResource?: string;
    channelId: string;
    appName?: string;
    appVersion?: string;
    appId?: string;
    logLevel?: LogLevel;
    customHeaders?: ICustomHeader;
}
export interface IHttpClientOptions {
    customHeaders?: ICustomHeader;
}
export interface ExtendedRTCIceServer extends RTCIceServer {
    type: string;
}
export type RequestApiOptions = {
    method: 'get' | 'post' | 'patch' | 'put' | 'delete';
    host: string;
    data?: any;
    version?: string;
    responseType?: ResponseType;
    contentType?: string;
    authToken?: string;
    logger?: any;
    noAuthHeader?: boolean;
    requestTimeout?: number;
    customHeaders?: ICustomHeader;
    maxAttempts?: number;
    signal?: AbortSignal;
};
// ...
export interface StreamingClientConnectOptions {
    /**
     * @deprecated since version 15.1.1. Please use maxConnectionAttempts instead
     */
    keepTryingOnFailure?: boolean;
    maxConnectionAttempts?: number;
    maxDelayBetweenConnectionAttempts?: number;
}